pyproject.toml
[tool.poetry]
name = "fastreact"
version = "0.1.0"
description = ""
authors = ["AndyPythonCode <andyarciniegas24@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.67.0"
aiofiles = "^0.7.0"
databases = "^0.4.3"
SQLAlchemy = "1.3.24"
aiosqlite = "^0.17.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.5"
pydantic = {extras = ["email"], version = "^1.8.2"}

[tool.poetry.dev-dependencies]
uvicorn = "^0.14.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
